fix: check current password before update/delete user

+++ app/controllers/userController.py
@@ -7,7 +7,7 @@ from fastapi.security import HTTPBearer
 from fastapi import Depends, HTTPException, status
 from passlib.context import CryptContext
 from models.user import UserModel
-from schemas.userSchema import RegisterUser, Login, UpdateUser
+from schemas.userSchema import ConfirmPassword, RegisterUser, Login, UpdateUser
 from dotenv import load_dotenv
 import os
 from fastapi.security import HTTPBearer
@@ -152,12 +152,16 @@ class UserController:
 
     def updateUser(userId: int, user: UpdateUser, db: Session):
         dbUserId = db.query(UserModel).filter(UserModel.id == userId).first()
+        if not verify(dbUserId.password, user.currentPass):
+            raise HTTPException(
+                status_code=status.HTTP_404_NOT_FOUND, detail=f"Incorrect password"
+            )
         if user.fullname is not None:
             dbUserId.fullname = user.fullname
         if user.email is not None:
             dbUserId.email = user.email
-        if user.password is not None:
-            dbUserId.password = bcrypt(user.password)
+        if user.newPassword is not None:
+            dbUserId.password = bcrypt(user.newPassword)
         if user.date_of_birth is not None:
             dbUserId.date_of_birth = user.date_of_birth
         # if user.role is not None:
@@ -167,7 +171,12 @@ class UserController:
         db.commit()
         return {"msg": "Updated"}
 
-    def deleteUser(userId: int, db: Session):
+    def deleteUser(userId: int, password: ConfirmPassword, db: Session):
+        dbUserId = db.query(UserModel).filter(UserModel.id == userId).first()
+        if not verify(dbUserId.password, password.currentPass):
+            raise HTTPException(
+                status_code=status.HTTP_404_NOT_FOUND, detail=f"Incorrect password"
+            )
         dbUserId = db.query(UserModel).filter(UserModel.id == userId).first()
         db.delete(dbUserId)
         db.commit()
